cbai/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
email-validator==2.2.0
httpx==0.28.1
redis==5.0.1
celery==5.3.4
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0 